+++ HISTORY.rst
@@ -46,6 +46,8 @@ History
     * Change sautils.JSONType to postgres dialect JSONB in all models (rudaporto).
     * Added new migration step to migrate all fields from json to jsonb (rudaporto).
     * Fixed localroles migration step: importing intenal_qa and internal_scout on the project level (rudaporto).
+    * Project: added new relationship to leadorders and also fix the orders relationship using the current_type in the join condition (rudaporto).
+    * Added new aggregator attribute to keep the total of leadorders and return this on the listing (rudaporto).
 
 2.1.35 (Unreleased)
 -------------------